requirements.txt
Flask==3.1.0
python-dotenv==1.1.0
numpy==2.1.1
pandas==2.2.3
opencv-python==4.11.0.86
torch==2.6.0
torchvision==0.21.0
xgboost==3.0.0
joblib==1.4.2
scikit-learn==1.6.1
scikit-optimize==0.10.2
sympy==1.13.1
PyYAML==6.0.2
ultralytics==8.3.104
werkzeug==3.1.3
tqdm==4.67.1
typing_extensions==4.13.1
setuptools==78.1.0
